app/requirements.txt
flask==2.3.3
gunicorn==20.1.0
google-cloud-firestore==2.13.1